Migrate Issuereport to repository (#3, #8 / #11)
This PR introduces the IssueReportProcessor.

The IssueReportProcessor takes a list of issues and the list of test,
annotated with the Issue-Annotation, and creates a report with these
information. The report shows how well issues are covered by tests.

At the moment the list of issues can only be provided as an CSV-file
and the report will be XML (generated using JAXB). Futher formats
are planed, but will be implemented independend. 

closes #3, #8 
PR: #11