Fix spotless, update build pipelines, remove moditec
Spotless threw errors on remove unused imports. This was
fixed the way in the main repo by using a java plugin.

Build pipeline was updated with current Java and JUnit versions.

Moditec and modular build was removed as the CVS depencendy does not support modular build. So the modular build would always fail.